+++ src/Runtime/Aggregator/LengthAggregator.php
@@ -3,18 +3,18 @@ declare(strict_types=1);
 
 namespace Remorhaz\JSON\Path\Runtime\Aggregator;
 
+use Remorhaz\JSON\Data\Iterator\ValueIteratorFactoryInterface;
 use function iterator_count;
 use Remorhaz\JSON\Data\Value\ArrayValueInterface;
 use Remorhaz\JSON\Path\Value\LiteralScalarValue;
 use Remorhaz\JSON\Data\Value\ValueInterface;
-use Remorhaz\JSON\Data\Iterator\ValueIteratorFactory;
 
 final class LengthAggregator implements ValueAggregatorInterface
 {
 
     private $valueIteratorFactory;
 
-    public function __construct(ValueIteratorFactory $valueIteratorFactory)
+    public function __construct(ValueIteratorFactoryInterface $valueIteratorFactory)
     {
         $this->valueIteratorFactory = $valueIteratorFactory;
     }
@@ -36,7 +36,7 @@ final class LengthAggregator implements ValueAggregatorInterface
 
         $arrayIterator = $this
             ->valueIteratorFactory
-            ->createArrayIterator($value->createIterator());
+            ->createArrayIterator($value->createEventIterator());
 
         return iterator_count($arrayIterator);
     }